server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
azure-storage-blob==12.9.0
pandas==1.3.0
python-dotenv==0.19.0